requirements.txt
Flask>=3.0.0,<4.0.0
pyodbc>=5.1.0
python-dotenv>=1.0.1
